requirements.txt
Flask==0.11.1
flask-cors==3.0.0
sqlalchemy==1.0.15
Flask-SQLAlchemy==2.1
psycopg2==2.6.2
python-jose==1.2.0
requests==2.11.1
cryptography==1.5.2